+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [1.1.0](https://github.com/ErezNagar/lease-calculator/compare/v1.0.1...v1.1.0) - 29/6/2020
+
+### Added
+
+- Acquisition and disposition fee support for 35 manufacturers:
+    calculate() now accepts a `make` field for calculating manufacturer-based fees.
+- getAcquisitionFeeValue(), returns acquisition fee value
+- getDispositionFeeValue(), returns disposition fee value
+
 ## [1.0.1](https://github.com/ErezNagar/lease-calculator/compare/v1.0.0...v1.0.1) - 27/4/2020
 
 ### Fixed